+++ lib/hit.dart
@@ -1,6 +1,6 @@
 library;
 
-export 'src/hit_defer.dart';
-export 'src/hit_layer.dart';
-export 'src/hit_link.dart';
-export 'src/hit_scope.dart';
+export 'src/hit_defer.dart' show Hit;
+export 'src/hit_layer.dart' show HitLayer;
+export 'src/hit_link.dart' show HitLink, HitDeferRegistration;
+export 'src/hit_scope.dart' show HitScope, HitScopeState;