+++ test/hit_overflow_perf_test.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:hit/hit.dart';
+import 'package:hit/src/hit_layer.dart';
 
 void main() {
   testWidgets('non-overflow HitLayer is not registered on HitLink', (